Condition new tests to only run in optimized builds.
