+++ soroban-env-host/src/test.rs
@@ -10,6 +10,8 @@ mod crypto;
 mod depth_limit;
 mod event;
 mod hostile;
+#[cfg(feature = "opt_build")]
+mod hostile_opt;
 mod invocation;
 mod ledger;
 mod lifecycle;
